C++ Hashing Project
C++ program to read in a large data file containing postal code and
location information for about 25,000 cities. It saves the city
information to an object as it is reading the file. We use a hash table
to store city objects and use a hashed city name as the hash key.
Collisions are handled by chaining using a linked list.